backend/requirements.txt
fastapi>=0.95.1,<0.96.0
starlette>=0.26.1,<0.27.0
uvicorn>=0.22.0,<0.23.0
python-multipart>=0.0.6,<0.1.0
pymongo>=4.7.0,<5.0.0
boto3>=1.34.0,<1.35.0
pdfminer.six>=20221105
pytest>=7.0.0,<8.0.0
python-dotenv
